services/surgery-service/requirements.txt
Django<=5.1
gunicorn==20.1.0
gevent==23.9.1
whitenoise==6.5.0
psycopg2-binary==2.9.5
dj-database-url==2.1.0
google-cloud-secret-manager==2.16.1
google-auth==2.23.0
django-environ==0.10.0
django-storages
google-cloud-storage
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
PyJWT==2.8.0
